Rules with transformations
Added tests for rules with transformation and a 
somewhat inacurate implementation of their parsing,
since there is still the context string in the match data.
Also removed the transformation component in the old test
since with an actuall rule for it in the parser there is no
empty strig matching for that rule.